requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==7.0.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
botocore==1.29.89
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-jsonfield-backport==1.0.5
frozenlist==1.3.3
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.4
jmespath==1.0.1
multidict==6.0.4
promise==2.3
psycopg-binary==3.1.8
psycopg2-binary==2.9.5
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
Rx==1.6.3
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
typing_extensions==4.5.0
urllib3==1.26.15
websockets==10.4
yarl==1.8.2
